PR 1099996 [Confidential] - [CLoS Setup] Device gets wrongly mapped when we delete and leaf and add a different new leaf
fix:
In case of ztpStage is true and this is a leaf and deployStatus changes to provision:
The device is going to be removed and the management ip will eventually expire and go back to DHCP pool.
We need to reset managementIp field for this device to None because soon another leaf will probably get the
same management ip address from DHCP. We want avoid the getDevices REST API returns
2 devices having the same management ip.
Note we don't do this when ztpStaged is false because all the management ip are pre-allocated
in that case and will never change.

PR: 1099996
Reviewed by: yunli